Updated the ALL config files to have the path to GLM. Added All config to scene.test. Some archi. changes
