+++ FireFlyRenderEngine/main.cpp
@@ -2,6 +2,7 @@
 #include <iostream>
 #include "Scene/Scene.h"
 #include "GLFWViewer/GLFWViewer.h"
+#include "../glm-0.9.9.7/vec3.hpp"
 
 int main(int argc, char* argv[])
 {
@@ -12,7 +13,8 @@ int main(int argc, char* argv[])
 	std::shared_ptr<Scene> scene = std::make_shared<Scene>();
 	std::string filePath = R"(C:\Users\rudra\Documents\Projects\FireflyRenderEngine\GPUPathTracer\SceneResources\cube.obj)";
 	scene->SetScreenWidthAndHeight(screenWidth, screenHeight);
-	scene->LoadScene(filePath);
+	scene->LoadOBJ(filePath);
+	scene->SetRasterCamera(glm::vec3(0.0f, 5.0f, 10.0f));
 
 	// Calls renderer
 	// gets film
