+++ GLFWViewer/GLFWViewer.h
@@ -34,7 +34,6 @@ public:
 
 	// This function is used to check & deal with any key press events
 	void ProcessKeyboardInput();
-	void ProcessMouseInput();
 
 	virtual ~GLFWViewer() override 
 	{
